♻️  [#156]; refactor: MapPage 방식으로 변경되기 전 사용되던 코드, 파일들 삭제  resolveModalFromWarnings 를 MapPage 에서 modalFlowLogic 으로 이동
