+++ pick-habju/src/pages/MapPage.tsx
@@ -19,26 +19,13 @@ import PartialReservationConfirmModal from '../components/Modal/Portion/PartialR
 import OneHourCallReservationNoticeModal from '../components/Modal/OneHour/OneHourCallReservationNoticeModal';
 import OneHourChatReservationNoticeModal from '../components/Modal/OneHour/OneHourChatReservationNoticeModal';
 import CallReservationNoticeModal from '../components/Modal/Call/CallReservationNoticeModal';
+import { resolveModalFromWarnings } from '../utils/modalFlowLogic';
 import type { ModalType } from '../utils/modalFlowLogic';
 import { formatAvailableTimeRange, formatTimeRangeForCard } from '../utils/availableTimeFormatter';
 import { getBookingUrl } from '../utils/bookingUrl';
-import type { PolicyWarning } from '../api/api.types';
 import BookModalStepper from '../components/Modal/Book/BookModal';
 
 const DEFAULT_MAP_ZOOM = 16;
-
-/**
- * policy_warnings 배열로 어떤 모달을 열지 결정.
- * partial 분기는 호출부에서 isPartialFilterActive로 먼저 처리한 뒤 이 함수를 호출.
- * 우선순위: chat_required_1h > call_required_1h > call_required_today > book
- */
-function resolveModalFromWarnings(warnings: PolicyWarning[]): Exclude<ModalType, 'partial'> {
-  const types = warnings.map((w) => w.type);
-  if (types.includes('chat_required_1h')) return 'oneHourChat';
-  if (types.includes('call_required_1h')) return 'oneHourCall';
-  if (types.includes('call_required_today')) return 'sameDayCall';
-  return 'book';
-}
 
 /**
  * 지도 기반 합주실 검색 페이지.
